Math: Inline function sofm_lut_sin_fixed_16b() for performance
This patch inlines the function sofm_lut_sin_fixed_16b() and
moves it to header file lut_trig.h. The lookup table is kept
in lut_trig.c and made global.

The DRC component use a lot the sine function (the fast lookup
tables version). The function seems to not get improvement from
HiFi intrinsics rewrite but making it inline improves DRC
performance in MTL platform by 0.54 MCPS, from 12.62 MCPS to
12.08 MCPS.

In Multiband-DRC the saving multiplies by number of bands, e.g.
1.58 MCPS saving with three bands.
